move keyboard back to core
