+++ src/interactive_pipe/graphical/qt_gui.py
@@ -33,7 +33,7 @@ if not PYQTVERSION:
 from interactive_pipe.core.control import Control
 from interactive_pipe.graphical.gui import InteractivePipeGUI, InteractivePipeWindow
 from interactive_pipe.graphical.qt_control import ControlFactory
-from interactive_pipe.graphical.keyboard import KeyboardControl
+from interactive_pipe.core.keyboard import KeyboardControl
 from interactive_pipe.headless.pipeline import HeadlessPipeline
 from typing import List
 import numpy as np
